#19 [refactor] 메모페이지 리팩토링
